save temp value
Added the ability to save the values of the tempcontrol. This could be improved at a later stage, but it is enough for testing with the NaLiControl
